mrb: add LoadEnv & rm cluster config & add Forker for apps & fix core & fix assert/logger & fix database module & fix knex & fix mysql free &

+++ src/databases/mysql/free/index.ts
@@ -1,6 +1,6 @@
 import { mdb } from '../config/knex.js';
 
 export async function freeMysql(): Promise<void> {
-  await mdb.destroy();
+  if (mdb != undefined) await mdb.destroy();
   echo('Success: Mysql now free');
 }
